feat: 호가 분리
